+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 ### Bug fixes and improvements
 
+- server: extend telemetry to handle multiple sources
 - console: allow schemas prefixed with `pg`, but not `pg_` (fix #8435)
 - console: add support for computed fields with session arg in permission builder (fix #8321)
-- server: extend telemetry to handle multiple sources
+- console: add GraphQL field customization for new database connections (root fields namespace, prefix, and suffix, and type names prefix and suffix)
 
 ## v2.8.0-beta.1